# Faster hash library xxh3