--- a/go.mod
+++ b/go.mod
@@ -2,4 +2,9 @@ module github.com/tidwall/hashmap
 
 go 1.18
 
-require github.com/cespare/xxhash/v2 v2.1.2
+require github.com/zeebo/xxh3 v1.0.2
+
+require (
+	github.com/klauspost/cpuid/v2 v2.1.1 // indirect
+	golang.org/x/sys v0.0.0-20220825204002-c680a09ffe64 // indirect
+)