## Proposed changes

- Add gitignore
- Lint README
- Update benchmark results in BENCH
- Use [xxh3](https://github.com/zeebo/xxh3), which I found to be significantly faster than xxhash in most cases as when compared to the standard library map. Benchmark comparison can be found below. Tested on AMD Ryzen 7 5800X, Manjaro Linux, Go 1.19. All test cases passed.

<table>
<tr>
<th>xxhash (old)</th>
<th>xxh3 (new)</th>
</tr>
<tr>
<td>

```shell
seed: 1661541371723767087

## STRING KEYS

-- tidwall --
set          100,000 ops     30ms      3,374,718/sec 
get          100,000 ops     19ms      5,195,199/sec 
reset        100,000 ops     16ms      6,210,687/sec 
scan              20 ops      9ms          2,106/sec 
delete       100,000 ops     17ms      6,034,829/sec 
memory     4,194,288 bytes                  41/entry 

-- stdlib --
set          100,000 ops     21ms      4,729,408/sec 
get          100,000 ops     11ms      9,132,903/sec 
reset        100,000 ops      8ms     13,307,516/sec 
scan              20 ops     25ms            791/sec 
delete       100,000 ops     10ms     10,505,883/sec 
memory     3,959,216 bytes                  39/entry


## INT KEYS

-- tidwall --
set          100,000 ops     15ms      6,758,532/sec 
get          100,000 ops      8ms     12,671,414/sec 
reset        100,000 ops      9ms     11,509,126/sec 
scan              20 ops      9ms          2,295/sec 
delete       100,000 ops      5ms     21,325,112/sec 
memory     3,143,368 bytes                  31/entry 

-- stdlib --
set          100,000 ops     12ms      8,096,538/sec 
get          100,000 ops      6ms     15,633,635/sec 
reset        100,000 ops      5ms     20,902,367/sec 
scan              20 ops     20ms            992/sec 
delete       100,000 ops      5ms     22,038,091/sec 
memory     2,765,832 bytes                  27/entry

seed: 1661541433491436408

## STRING KEYS

-- tidwall --
set        1,000,000 ops    286ms      3,493,008/sec 
get        1,000,000 ops    215ms      4,643,501/sec 
reset      1,000,000 ops    218ms      4,590,131/sec 
scan              20 ops    135ms            148/sec 
delete     1,000,000 ops    247ms      4,052,964/sec 
memory    67,108,848 bytes                  67/entry 

-- stdlib --
set        1,000,000 ops    328ms      3,052,252/sec 
get        1,000,000 ops    120ms      8,355,899/sec 
reset      1,000,000 ops    123ms      8,134,754/sec 
scan              20 ops    155ms            129/sec 
delete     1,000,000 ops    138ms      7,264,480/sec 
memory    57,931,472 bytes                  57/entry


## INT KEYS

-- tidwall --
set        1,000,000 ops    151ms      6,612,492/sec 
get        1,000,000 ops    113ms      8,845,072/sec 
reset      1,000,000 ops    110ms      9,114,482/sec 
scan              20 ops    138ms            145/sec 
delete     1,000,000 ops    109ms      9,197,197/sec 
memory    50,329,288 bytes                  50/entry 

-- stdlib --
set        1,000,000 ops    169ms      5,920,471/sec 
get        1,000,000 ops     75ms     13,293,602/sec 
reset      1,000,000 ops     60ms     16,693,933/sec 
scan              20 ops    155ms            129/sec 
delete     1,000,000 ops     69ms     14,499,619/sec 
memory    40,146,664 bytes                  40/entry

seed: 1661541461974249082

## STRING KEYS

-- tidwall --
set       10,000,000 ops   3165ms      3,159,250/sec 
get       10,000,000 ops   2589ms      3,862,272/sec 
reset     10,000,000 ops   2636ms      3,792,936/sec 
scan              20 ops   1031ms             19/sec 
delete    10,000,000 ops   2849ms      3,509,885/sec 
memory   536,870,904 bytes                  53/entry 

-- stdlib --
set       10,000,000 ops   3519ms      2,842,029/sec 
get       10,000,000 ops   1464ms      6,831,421/sec 
reset     10,000,000 ops   1437ms      6,960,506/sec 
scan              20 ops   1705ms             11/sec 
delete    10,000,000 ops   1557ms      6,422,395/sec 
memory   463,468,256 bytes                  46/entry 


## INT KEYS

-- tidwall --
set       10,000,000 ops   1691ms      5,915,152/sec 
get       10,000,000 ops   1326ms      7,544,039/sec 
reset     10,000,000 ops   1373ms      7,285,648/sec 
scan              20 ops   1073ms             18/sec 
delete    10,000,000 ops   1465ms      6,828,060/sec 
memory   402,650,824 bytes                  40/entry 

-- stdlib --
set       10,000,000 ops   1625ms      6,155,027/sec 
get       10,000,000 ops    745ms     13,430,294/sec 
reset     10,000,000 ops    828ms     12,074,512/sec 
scan              20 ops   1578ms             12/sec 
delete    10,000,000 ops    867ms     11,535,263/sec 
memory   321,976,040 bytes                  32/entry
```

</td>
<td>

```shell
seed: 1661541547938461762

## STRING KEYS

-- tidwall --
set          100,000 ops     24ms      4,149,762/sec 
get          100,000 ops     15ms      6,536,763/sec 
reset        100,000 ops     10ms      9,774,736/sec 
scan              20 ops     10ms          1,986/sec 
delete       100,000 ops      7ms     14,222,727/sec 
memory     4,194,288 bytes                  41/entry 

-- stdlib --
set          100,000 ops     17ms      5,818,370/sec 
get          100,000 ops      9ms     11,519,833/sec 
reset        100,000 ops      6ms     15,533,085/sec 
scan              20 ops     20ms            990/sec 
delete       100,000 ops     10ms     10,442,394/sec 
memory     3,961,296 bytes                  39/entry 


## INT KEYS

-- tidwall --
set          100,000 ops     12ms      8,150,474/sec 
get          100,000 ops      6ms     15,637,994/sec 
reset        100,000 ops      6ms     15,584,477/sec 
scan              20 ops      9ms          2,128/sec 
delete       100,000 ops      5ms     19,495,472/sec 
memory     3,143,352 bytes                  31/entry 

-- stdlib --
set          100,000 ops     10ms      9,584,498/sec 
get          100,000 ops      6ms     16,771,648/sec 
reset        100,000 ops      6ms     17,205,739/sec 
scan              20 ops     20ms          1,010/sec 
delete       100,000 ops      4ms     23,833,010/sec 
memory     2,766,120 bytes                  27/entry

seed: 1661541569345505845

## STRING KEYS

-- tidwall --
set        1,000,000 ops    221ms      4,517,183/sec 
get        1,000,000 ops    128ms      7,838,039/sec 
reset      1,000,000 ops    126ms      7,951,959/sec 
scan              20 ops    137ms            145/sec 
delete     1,000,000 ops    145ms      6,917,694/sec 
memory    67,108,848 bytes                  67/entry 

-- stdlib --
set        1,000,000 ops    309ms      3,232,441/sec 
get        1,000,000 ops    124ms      8,087,531/sec 
reset      1,000,000 ops    125ms      8,005,564/sec 
scan              20 ops    165ms            120/sec 
delete     1,000,000 ops    142ms      7,039,927/sec 
memory    57,931,568 bytes                  57/entry 


## INT KEYS

-- tidwall --
set        1,000,000 ops    124ms      8,067,036/sec 
get        1,000,000 ops     72ms     13,980,887/sec 
reset      1,000,000 ops     64ms     15,596,642/sec 
scan              20 ops    137ms            145/sec 
delete     1,000,000 ops     65ms     15,325,288/sec 
memory    50,329,272 bytes                  50/entry 

-- stdlib --
set        1,000,000 ops    126ms      7,921,841/sec 
get        1,000,000 ops     63ms     15,880,942/sec 
reset      1,000,000 ops     62ms     16,024,580/sec 
scan              20 ops    154ms            130/sec 
delete     1,000,000 ops     73ms     13,747,306/sec 
memory    40,146,856 bytes                  40/entry

seed: 1661541603926099681

## STRING KEYS

-- tidwall --
set       10,000,000 ops   2496ms      4,006,428/sec 
get       10,000,000 ops   1404ms      7,122,732/sec 
reset     10,000,000 ops   1460ms      6,848,696/sec 
scan              20 ops   1031ms             19/sec 
delete    10,000,000 ops   1710ms      5,848,611/sec 
memory   536,870,896 bytes                  53/entry 

-- stdlib --
set       10,000,000 ops   3678ms      2,719,207/sec 
get       10,000,000 ops   1476ms      6,775,262/sec 
reset     10,000,000 ops   1466ms      6,820,987/sec 
scan              20 ops   1770ms             11/sec 
delete    10,000,000 ops   1594ms      6,274,049/sec 
memory   463,468,240 bytes                  46/entry 


## INT KEYS

-- tidwall --
set       10,000,000 ops   1399ms      7,149,446/sec 
get       10,000,000 ops    720ms     13,888,579/sec 
reset     10,000,000 ops    782ms     12,793,742/sec 
scan              20 ops   1075ms             18/sec 
delete    10,000,000 ops    898ms     11,131,669/sec 
memory   402,650,808 bytes                  40/entry 

-- stdlib --
set       10,000,000 ops   1639ms      6,100,070/sec 
get       10,000,000 ops    770ms     12,988,335/sec 
reset     10,000,000 ops    852ms     11,743,657/sec 
scan              20 ops   1591ms             12/sec 
delete    10,000,000 ops    887ms     11,279,646/sec 
memory   321,976,040 bytes                  32/entry
```

</td>
</tr>
</table>