@tidwall 2022-08-26
Hi, I'm currently running a macbook with an M1 Max. I'm seeing about the same performance with the current xxhash library, maybe xxh3 is smidge faster, but it's a close call. But I will have to test on an AMD machine to see what you're seeing. The XXH3 algo looks promising.