@elliotwutingfeng 2022-08-27
Hi, I've just ran some benchmarks on a Raspberry Pi 3. There appears to be little change except for INT keys being slightly faster at 1000_000 ops for "get" and "reset".
---

Tested on
- Raspberry Pi 3 Model B Rev 1.2
- Linux raspberrypi 5.15.38-v7+ #1553 SMP Wed May 11 16:46:03 BST 2022 armv7l GNU/Linux
- Go 1.19
- Not enough RAM to test 10_000_000 ops :smile:

<table>
<tr>
<th>xxhash (old)</th>
<th>xxh3 (new)</th>
</tr>
<tr>
<td>

```shell
seed: 1661618565010438196

## STRING KEYS

-- tidwall --
set          100,000 ops    320ms        312,621/sec 
get          100,000 ops    117ms        855,457/sec 
reset        100,000 ops    148ms        675,236/sec 
scan              20 ops    166ms            120/sec 
delete       100,000 ops    166ms        601,621/sec 
memory     2,621,432 bytes                  26/entry 

-- stdlib --
set          100,000 ops    208ms        480,091/sec 
get          100,000 ops     87ms      1,148,189/sec 
reset        100,000 ops    110ms        908,627/sec 
scan              20 ops    444ms             45/sec 
delete       100,000 ops    122ms        822,820/sec 
memory     2,061,032 bytes                  20/entry 


## INT KEYS

-- tidwall --
set          100,000 ops    149ms        669,570/sec 
get          100,000 ops     66ms      1,510,610/sec 
reset        100,000 ops     87ms      1,146,976/sec 
scan              20 ops    160ms            125/sec 
delete       100,000 ops    103ms        970,294/sec 
memory     2,096,440 bytes                  20/entry 

-- stdlib --
set          100,000 ops    147ms        681,763/sec 
get          100,000 ops     60ms      1,679,055/sec 
reset        100,000 ops     71ms      1,401,446/sec 
scan              20 ops    335ms             59/sec 
delete       100,000 ops     43ms      2,317,131/sec 
memory     1,471,208 bytes                  14/entry 

seed: 1661618598661568236

## STRING KEYS

-- tidwall --
set        1,000,000 ops   3800ms        263,132/sec 
get        1,000,000 ops   1115ms        896,597/sec 
reset      1,000,000 ops    830ms      1,205,170/sec 
scan              20 ops   1105ms             18/sec 
delete     1,000,000 ops   1031ms        970,333/sec 
memory    41,943,032 bytes                  41/entry 

-- stdlib --
set        1,000,000 ops   3061ms        326,677/sec 
get        1,000,000 ops   1066ms        937,729/sec 
reset      1,000,000 ops    761ms      1,314,061/sec 
scan              20 ops   2749ms              7/sec 
delete     1,000,000 ops   1250ms        800,221/sec 
memory    30,080,264 bytes                  30/entry 


## INT KEYS

-- tidwall --
set        1,000,000 ops   1893ms        528,293/sec 
get        1,000,000 ops    654ms      1,529,229/sec 
reset      1,000,000 ops    803ms      1,244,899/sec 
scan              20 ops   1182ms             16/sec 
delete     1,000,000 ops    646ms      1,547,128/sec 
memory    33,553,656 bytes                  33/entry 

-- stdlib --
set        1,000,000 ops   2050ms        487,803/sec 
get        1,000,000 ops    391ms      2,556,562/sec 
reset      1,000,000 ops    446ms      2,241,969/sec 
scan              20 ops   2323ms              8/sec 
delete     1,000,000 ops    477ms      2,098,453/sec 
memory    21,190,136 bytes                  21/entry
```

</td>
<td>

```shell
seed: 1661618384777358290

## STRING KEYS

-- tidwall --
set          100,000 ops    315ms        317,186/sec 
get          100,000 ops    117ms        855,893/sec 
reset        100,000 ops    148ms        677,523/sec 
scan              20 ops    165ms            121/sec 
delete       100,000 ops    165ms        604,324/sec 
memory     2,621,432 bytes                  26/entry 

-- stdlib --
set          100,000 ops    224ms        445,624/sec 
get          100,000 ops    104ms        962,860/sec 
reset        100,000 ops    111ms        902,644/sec 
scan              20 ops    454ms             44/sec 
delete       100,000 ops    122ms        816,491/sec 
memory     2,069,656 bytes                  20/entry 


## INT KEYS

-- tidwall --
set          100,000 ops    151ms        663,760/sec 
get          100,000 ops     67ms      1,498,628/sec 
reset        100,000 ops     86ms      1,158,541/sec 
scan              20 ops    159ms            125/sec 
delete       100,000 ops    104ms        962,713/sec 
memory     2,096,440 bytes                  20/entry 

-- stdlib --
set          100,000 ops    147ms        681,609/sec 
get          100,000 ops     60ms      1,671,421/sec 
reset        100,000 ops     71ms      1,408,841/sec 
scan              20 ops    216ms             92/sec 
delete       100,000 ops     43ms      2,312,391/sec 
memory     1,464,328 bytes                  14/entry

seed: 1661618471226468475

## STRING KEYS

-- tidwall --
set        1,000,000 ops   3857ms        259,298/sec 
get        1,000,000 ops   1166ms        857,664/sec 
reset      1,000,000 ops    839ms      1,192,493/sec 
scan              20 ops   1107ms             18/sec 
delete     1,000,000 ops   1039ms        962,274/sec 
memory    41,943,032 bytes                  41/entry 

-- stdlib --
set        1,000,000 ops   3161ms        316,340/sec 
get        1,000,000 ops   1008ms        992,348/sec 
reset      1,000,000 ops    718ms      1,393,228/sec 
scan              20 ops   4158ms              4/sec 
delete     1,000,000 ops    762ms      1,311,981/sec 
memory    30,080,264 bytes                  30/entry 


## INT KEYS

-- tidwall --
set        1,000,000 ops   1865ms        536,284/sec 
get        1,000,000 ops    385ms      2,600,720/sec 
reset      1,000,000 ops    462ms      2,162,468/sec 
scan              20 ops   1057ms             18/sec 
delete     1,000,000 ops    647ms      1,544,881/sec 
memory    33,553,720 bytes                  33/entry 

-- stdlib --
set        1,000,000 ops   1668ms        599,500/sec 
get        1,000,000 ops    657ms      1,521,335/sec 
reset      1,000,000 ops    447ms      2,238,490/sec 
scan              20 ops   3924ms              5/sec 
delete     1,000,000 ops    501ms      1,995,880/sec 
memory    21,189,144 bytes                  21/entry
```

</td>
</tr>
</table>